move sendEmail to another file to avoid circular dependency
