+++ app/libs/email.js
@@ -1,31 +1,8 @@
-const nodemailer = require('nodemailer');
 const CONFIG = require('../config');
 const db = require('../models');
 const {addJobToEmailQueue} = require('../queue');
 
-const {email, password, domain, ehost} = CONFIG.get('email');
-
-const sendEmail = async (subject, text, toEmail) => {
-  const transporter = nodemailer.createTransport({
-    host: ehost,
-    auth: {
-      user: email,
-      pass: password,
-    },
-    tls: {
-      rejectUnauthorized: false,
-    },
-  });
-
-  const mailOptions = {
-    from: `${email}${domain}`,
-    to: toEmail,
-    subject,
-    text,
-  };
-
-  return transporter.sendMail(mailOptions);
-};
+const {email, domain} = CONFIG.get('email');
 
 const getEmailList = async (triggers) => {
   const notifs = await db.models.notification.scope('extended').findAll({
@@ -69,4 +46,4 @@ const notifyUsers = async (subject, text, triggers) => {
   }
 };
 
-module.exports = {sendEmail, getEmailList, notifyUsers};
+module.exports = {getEmailList, notifyUsers};
