+++ app/queue.js
@@ -4,7 +4,7 @@ const {graphkbGetReadonlyGroupId, graphkbAddUser} = require('./api/graphkb');
 const conf = require('./config');
 const createReport = require('./libs/createReport');
 const db = require('./models');
-const {sendEmail} = require('./libs/email');
+const {sendEmail} = require('./libs/mailer');
 
 const {host, port, enableQueue} = conf.get('redis_queue');
 const logger = require('./log');
