+++ app/routes/report/reportUser.js
@@ -153,7 +153,6 @@ router.route('/')
             projectId: reportProject.project_id,
           },
         );
-        logger.info('Email sent successfully');
       } catch (error) {
         logger.error(`Email not sent successfully: ${error}`);
       }
